backend/requirements.txt
bcrypt==3.1.7
billiard==3.6.2.0
cachetools==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
Django==2.1.4
django-appconf==1.0.2
django-celery==3.2.2
django-cors-headers==2.4.0
django-crequest==2018.5.11
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-imagekit==4.0.2
django-import-export==1.1.0
django-ipware==2.1.0
django-redis-cache==2.1.0
django-rest-swagger==2.2.0
django-reversion==3.0.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
et-xmlfile==1.0.1
flower==0.9.3
future==0.17.1
gevent==1.4.0
google-auth==1.11.2
gunicorn==19.9.0
httplib2==0.12.0
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.5.0
ldap3==2.5.2
mysqlclient==1.3.14
oauthlib==3.1.0
paramiko==2.7.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyotp==2.2.7
python-dotenv==0.10.3
python-ldap==3.1.0
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
qrcode==6.1
regex==2020.2.20
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
tablib==0.12.1
txaio==20.1.1
tzlocal==2.0.0
unicodecsv==0.14.1
urllib3==1.24.3
xlwt==1.3.0
zmail==0.2.3
celery==4.3.0
redis==3.2.0
anyjson==0.3.3
pilkit==2.0
Pillow==5.3.0